helios/main.py
"""
Helios Trading Analysis Toolkit
Main application module
"""

import os
import sys
import argparse
import pandas as pd
import numpy as np
from dotenv import load_dotenv
from pathlib import Path


def setup_environment():
    """Load environment variables and setup configuration"""
    load_dotenv()
    
    # Example: Get API keys or config from environment
    api_key = os.getenv('HELIOS_API_KEY')
    data_path = os.getenv('HELIOS_DATA_PATH', './data')
    
    return {
        'api_key': api_key,
        'data_path': Path(data_path)
    }


def analyze_data(df):
    """Example analysis function"""
    print("\nData Analysis Summary:")
    print(f"Shape: {df.shape}")
    print(f"\nColumns: {list(df.columns)}")
    print(f"\nBasic Statistics:")
    print(df.describe())
    
    return df


def main():
    """Main entry point"""
    parser = argparse.ArgumentParser(description='Helios Trading Analysis Toolkit')
    parser.add_argument('--data', '-d', help='Path to data file')
    parser.add_argument('--analyze', '-a', action='store_true', 
                       help='Run analysis on data')
    parser.add_argument('--version', '-v', action='store_true',
                       help='Show version')
    
    args = parser.parse_args()
    
    if args.version:
        print("Helios v0.1.0")
        return 0
    
    # Setup environment
    config = setup_environment()
    print("Helios Trading Analysis Toolkit")
    print("=" * 40)
    
    # Example: Load and analyze data if provided
    if args.data:
        try:
            print(f"\nLoading data from: {args.data}")
            
            # Determine file type and load accordingly
            if args.data.endswith('.csv'):
                df = pd.read_csv(args.data)
            elif args.data.endswith('.parquet'):
                df = pd.read_parquet(args.data)
            else:
                print(f"Unsupported file format: {args.data}")
                return 1
            
            if args.analyze:
                analyze_data(df)
                
        except Exception as e:
            print(f"Error loading data: {e}")
            return 1
    else:
        print("\nNo data file specified. Use --data to load a file.")
        print("Example: python -m helios --data mydata.csv --analyze")
    
    return 0


if __name__ == "__main__":
    sys.exit(main())
