helios/pyproject.toml
[project]
name = "helios"
version = "0.1.0"
description = "Helios trading analysis toolkit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "jupyter>=1.0.0",
]
